refactor(handlers): extract feature-cache.ts from intent.ts (Stage B)
Move the cachedFeatures/cachedDiscovery module state + 5 accessors to one
module. intent.ts reads the state via ESM live bindings (60 reads unchanged);
the single probe mutation routes through setCachedFeatures(). Accessors
re-exported for back-compat. 3672 tests green, A1 zero-diff.
